+++ src/PowerShell/Microsoft.WinGet.Client.Engine/Commands/Base/BaseCommand.cs
@@ -1,0 +1,30 @@
+﻿// -----------------------------------------------------------------------------
+// <copyright file="BaseCommand.cs" company="Microsoft Corporation">
+//     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
+// </copyright>
+// -----------------------------------------------------------------------------
+
+namespace Microsoft.WinGet.Client.Engine.Commands.Base
+{
+    using System.Management.Automation;
+    using Microsoft.WinGet.Client.Engine.Common;
+    using Microsoft.WinGet.Client.Engine.Exceptions;
+
+    /// <summary>
+    /// Base class for all Cmdlets.
+    /// </summary>
+    public abstract class BaseCommand
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BaseCommand"/> class.
+        /// </summary>
+        public BaseCommand()
+            : base()
+        {
+            if (Utilities.ExecutingAsSystem)
+            {
+                throw new ExecuteAsSystemException();
+            }
+        }
+    }
+}
