+++ src/PowerShell/Microsoft.WinGet.Client.Engine/Common/ErrorCode.cs
@@ -4,7 +4,7 @@
 // </copyright>
 // -----------------------------------------------------------------------------
 
-namespace Microsoft.WinGet.Client.Common
+namespace Microsoft.WinGet.Client.Engine.Common
 {
     /// <summary>
     /// Error code constants.
